+++ templates/festival.html
@@ -60,15 +60,15 @@
 <div class="container">
 
     <!-- Page Heading/Breadcrumbs -->
-    <h1 class="mt-4 mb-3">축제/행사/관광지
-        <small>Subheading</small>
+    <h1 class="mt-4 mb-3">축제/행사
+        <small></small>
     </h1>
 
     <ol class="breadcrumb">
         <li class="breadcrumb-item">
             <a href="index.html">Home</a>
         </li>
-        <li class="breadcrumb-item active">축제/행사/관광지</li>
+        <li class="breadcrumb-item active">축제/행사</li>
     </ol>
 
     <!-- Project One -->
