+++ templates/index.html
@@ -113,9 +113,9 @@
       </div>
       <div class="col-lg-4 mb-4">
         <div class="card h-100">
-          <h4 class="card-header">축제/행사/관광지</h4>
+          <h4 class="card-header">축제/행사</h4>
           <div class="card-body">
-            <p class="card-text">Lorem ipsum dolor sit amet, consectetur adipisicing elit. Reiciendis ipsam eos, nam perspiciatis natus commodi similique totam consectetur praesentium molestiae atque exercitationem ut consequuntur, sed eveniet, magni nostrum sint fuga.</p>
+            <p class="card-text">아이들과 함께 즐길 수 있는 주변 축제/행사 정보를 제공합니다.</p>
           </div>
           <div class="card-footer">
             <a href="festival.html" class="btn btn-primary">Learn More</a>
